# docs/swagger.yaml
basePath: /
definitions:
  hauslet_internal_modules_auth_domain.BackupCodesResult:
    properties:
      codes:
        description: Plain text codes (shown once)
        items:
          type: string
        type: array
    type: object
...
  hauslet_internal_modules_auth_domain.ErrorResponse:
    properties:
      error:
        type: string
      field:
        type: string
      message:
        type: string
    type: object
  hauslet_internal_modules_auth_domain.RegisterRequest:
    properties:
      birth_date:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  hauslet_internal_modules_auth_domain.RegisterResponse:
    properties:
      email:
        type: string
      id:
        type: string
      message:
        type: string
      name:
        type: string
    type: object
  hauslet_internal_modules_auth_domain.SessionResponse:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      ip:
        type: string
      provider:
        type: string
      user_agent:
        type: string
    type: object
  hauslet_internal_modules_auth_domain.SetupResponse:
    properties:
      method:
        $ref: '#/definitions/hauslet_internal_modules_auth_domain.TwoFactorMethod'
      qr_code_image:
        description: Base64 data URL for QR code image
        type: string
      qr_code_url:
        description: otpauth:// URL (for authenticator)
        type: string
      secret:
        description: Only for authenticator (shown once)
        type: string
    type: object
  hauslet_internal_modules_auth_domain.TwoFactorMethod:
    enum:
    - email
    - sms
    - authenticator
    type: string
    x-enum-varnames:
    - TwoFactorEmail
    - TwoFactorSMS
    - TwoFactorAuthenticator
  hauslet_internal_modules_auth_domain.TwoFactorStatus:
    properties:
      backup_codes_left:
        type: integer
      enabled:
        type: boolean
      method:
        $ref: '#/definitions/hauslet_internal_modules_auth_domain.TwoFactorMethod'
    type: object
  hauslet_internal_modules_auth_domain.UpdateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
    type: object
  hauslet_internal_modules_auth_domain.UserIdentityResponse:
    properties:
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      provider:
        type: string
    type: object
  hauslet_internal_modules_auth_domain.UserResponse:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  hauslet_internal_modules_auth_domain.Verify2FALoginRequest:
    properties:
      code:
        type: string
      temp_token:
        type: string
    type: object
...
  internal_modules_auth_port_http.ForgotPasswordRequest:
    properties:
      email:
        type: string
    type: object
  internal_modules_auth_port_http.PasswordlessLoginRequest:
    properties:
      code:
        type: string
      email:
        type: string
    type: object
  internal_modules_auth_port_http.ResendOTPRequest:
    properties:
      email:
        type: string
    type: object
  internal_modules_auth_port_http.ResetPasswordRequest:
    properties:
      email:
        type: string
      new_password:
        type: string
      token:
        type: string
    type: object
  internal_modules_auth_port_http.Setup2FARequest:
    properties:
      method:
        description: '"email", "sms", "authenticator"'
        type: string
      phone_number:
        description: Required for SMS method
        type: string
    type: object
  internal_modules_auth_port_http.Verify2FARequest:
    properties:
      code:
        type: string
    type: object
  internal_modules_auth_port_http.VerifyEmailRequest:
    properties:
      code:
        type: string
      email:
        type: string
    type: object
...
host: https://dev-api.hauslet.com
info:
  contact:
    email: support@hauslet.com
    name: API Support
    url: http://www.swagger.io/support
  description: Hauslet Services API Server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Hauslet API
  version: "1.0"
paths:
  /2fa/backup-codes:
    post:
      consumes:
      - application/json
      description: Generates new backup codes after verifying current 2FA code
      parameters:
      - description: Current 2FA code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth_port_http.Verify2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.BackupCodesResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Regenerate backup codes
      tags:
      - 2FA
  /2fa/disable:
    post:
      consumes:
      - application/json
      description: Disables 2FA after verifying current code
      parameters:
      - description: Current 2FA code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth_port_http.Verify2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Disable 2FA
      tags:
      - 2FA
  /2fa/send-code:
    post:
      description: Sends a new verification code for Email/SMS 2FA
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Send 2FA code
      tags:
      - 2FA
  /2fa/setup:
    post:
      consumes:
      - application/json
      description: Starts 2FA setup with the selected method
      parameters:
      - description: Setup request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth_port_http.Setup2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.SetupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Initiate 2FA setup
      tags:
      - 2FA
  /2fa/setup/complete:
    post:
      consumes:
      - application/json
      description: Verifies the code and enables 2FA
      parameters:
      - description: Verification code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth_port_http.Verify2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.BackupCodesResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Complete 2FA setup
      tags:
      - 2FA
  /2fa/status:
    get:
      description: Returns whether 2FA is enabled and which method
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.TwoFactorStatus'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get 2FA status
      tags:
      - 2FA
  /2fa/verify:
    post:
      consumes:
      - application/json
      description: Verifies the 2FA code to complete login
      parameters:
      - description: Verification code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth_port_http.Verify2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
      summary: Verify 2FA code
      tags:
      - 2FA
...
  /auth/2fa/resend:
    post:
      consumes:
      - application/json
      description: Resends the verification code for pending 2FA login
      parameters:
      - description: Temp token
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
      summary: Resend 2FA code
      tags:
      - 2FA
  /auth/2fa/verify:
    post:
      consumes:
      - application/json
      description: Verifies the 2FA code and issues a full JWT token
      parameters:
      - description: Temp token and verification code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hauslet_internal_modules_auth_domain.Verify2FALoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
      summary: Complete 2FA login
      tags:
      - 2FA
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send a password reset email to the user
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth_port_http.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
      summary: Request password reset
      tags:
      - auth
  /auth/link/{provider}:
    get:
      description: Start OAuth flow to link a provider (google) to authenticated user
        account
      parameters:
      - description: Provider name (google)
        in: path
        name: provider
        required: true
        type: string
      - description: 'URI to redirect after linking (default: /settings)'
        in: query
        name: redirect_uri
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to OAuth provider
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Initiate identity linking
      tags:
      - user
...
  /auth/passwordless/verify:
    post:
      consumes:
      - application/json
      description: Verify the 6-digit code sent to email and issue a JWT session
      parameters:
      - description: Email and verification code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth_port_http.PasswordlessLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
      summary: Verify passwordless login code
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email and password
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hauslet_internal_modules_auth_domain.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /auth/resend-otp:
    post:
      consumes:
      - application/json
      description: Resend verification code to user's email
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth_port_http.ResendOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
      summary: Resend OTP code
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password using the token sent via email
      parameters:
      - description: Reset details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth_port_http.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
      summary: Reset password
      tags:
      - auth
...
  /auth/verify-email:
    post:
      consumes:
      - application/json
      description: Verify user's email address using OTP code
      parameters:
      - description: Verification details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth_port_http.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
      summary: Verify email address
      tags:
      - auth
...
  /me:
    get:
      description: Returns the authenticated user's information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
        "403":
          description: 2FA Required
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update user's name or email
      parameters:
      - description: Update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hauslet_internal_modules_auth_domain.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - user
  /me/identities:
    get:
      description: List all authentication methods (OAuth providers and password)
        linked to the user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/hauslet_internal_modules_auth_domain.UserIdentityResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user identities
      tags:
      - user
...
  /me/sessions:
    delete:
      description: Revoke all sessions for the authenticated user (useful for "logout
        everywhere")
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Revoke all sessions
      tags:
      - session
    get:
      description: List all active sessions for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/hauslet_internal_modules_auth_domain.SessionResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hauslet_internal_modules_auth_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user sessions
      tags:
      - session
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"